circular in dicom loader
