+++ packages/dicomImageLoader/src/imageLoader/wadors/metaDataManager.ts
@@ -1,16 +1,47 @@
 import type { WADORSMetaData } from '../../types';
 import imageIdToURI from '../imageIdToURI';
 import { combineFrameInstance } from './combineFrameInstance';
-import multiframeMetadata from './retrieveMultiframeMetadata';
 
 let metadataByImageURI = [];
 let multiframeMetadataByImageURI = {};
+
+import getValue from './metaData/getValue';
+
+// get metadata information for the first frame
+function _retrieveMultiframeMetadata(imageURI) {
+  const lastSlashIdx = imageURI.indexOf('/frames/') + 8;
+  // imageid string without frame number
+  const imageIdFrameless = imageURI.slice(0, lastSlashIdx);
+  // calculating frame number
+  const frame = parseInt(imageURI.slice(lastSlashIdx), 10);
+  // retrieving the frame 1 that contains multiframe information
+
+  const metadata = metadataByImageURI[`${imageIdFrameless}1`];
+
+  return {
+    metadata,
+    frame,
+  };
+}
+
+function retrieveMultiframeMetadata(imageId) {
+  const imageURI = imageIdToURI(imageId);
+
+  return _retrieveMultiframeMetadata(imageURI);
+}
+
+function isMultiframe(metadata) {
+  // Checks if dicomTag NumberOf Frames exists and it is greater than one
+  const numberOfFrames = getValue<number>(metadata['00280008']);
+
+  return numberOfFrames && numberOfFrames > 1;
+}
 
 function add(imageId: string, metadata: WADORSMetaData) {
   const imageURI = imageIdToURI(imageId);
 
   Object.defineProperty(metadata, 'isMultiframe', {
-    value: multiframeMetadata.isMultiframe(metadata),
+    value: isMultiframe(metadata),
     enumerable: false,
   });
 
@@ -39,8 +70,7 @@ function get(imageId: string): WADORSMetaData {
   }
 
   // Try to get the metadata for a specific frame of a multiframe image
-  const retrievedMetadata =
-    multiframeMetadata._retrieveMultiframeMetadata(imageURI);
+  const retrievedMetadata = retrieveMultiframeMetadata(imageURI);
 
   if (!retrievedMetadata || !retrievedMetadata.metadata) {
     return;
@@ -71,7 +101,7 @@ function purge() {
   multiframeMetadataByImageURI = {};
 }
 
-export { metadataByImageURI };
+export { metadataByImageURI, isMultiframe, retrieveMultiframeMetadata };
 
 export default {
   add,
